move native managers to subdirectory
