+++ bridge-client/src/iosMain/kotlin/org/sagebionetworks/bridge/kmm/shared/managers/NativeHistoryManager.kt
@@ -1,9 +1,7 @@
-package org.sagebionetworks.bridge.kmm.shared
+package org.sagebionetworks.bridge.kmm.shared.managers
 
 import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.cancel
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
